Pass options to runTopology.
