+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nats-topology-runner",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "description": "Run a topology created with topology-runner using nats-jobs",
   "author": "GovSpend",
   "main": "dist/index.js",
